final changes => add_to_cart + check_out

+++ routes/web.php
@@ -6,6 +6,7 @@ Route::get('/','indexController@index');
 
 // About Controller
 
+
 Route::get('/about','aboutController@index');
 
 // Contact Controller
@@ -17,34 +18,35 @@ Route::post('/contact/send','contactController@send')->name('sendMessage');
 
 // Route::get('/menu','menuController@index'); 
 
-// Reservation Controller
-
-Route::get('/reservation','reservationController@index'); 
-Route::POST('/reservation/store','reservationController@store')->name('reservation.store');
+// Reservation Control
+Route::get('/reservation','reservationController@index')->middleware(['auth', 'role:customer']); 
+Route::POST('/reservation/store','reservationController@store')->name('reservation.store')->middleware(['auth', 'role:customer']);
 
 
 
 // Category Controller
-
-Route::get('/categories/create','categoryController@create')->middleware('auth'); 
-Route::POST('/categories','categoryController@store')->middleware('auth'); 
+Route::get('/categories/create','categoryController@create')->middleware(['auth', 'role:admin']); 
+Route::POST('/categories','categoryController@store')->middleware(['auth', 'role:admin']); 
 
 
 // Meal Controller
 Route::get('/meals','mealController@index'); 
-
-Route::get('/meals/create','mealController@create')->middleware('auth'); 
-Route::POST('/meals/store','mealController@store')->middleware('auth'); 
-Route::get('/meals/{id}/edit' ,'mealController@edit')->middleware('auth');
-Route::PATCH('/meals/{id}' ,'mealController@update')->middleware('auth');
-Route::delete('/meals/{id}'   ,'mealController@destroy')->middleware('auth');
-
+Route::get('/meals/create','mealController@create')->middleware(['auth', 'role:admin']); 
+Route::POST('/meals/store','mealController@store')->middleware(['auth', 'role:admin']); 
+Route::get('/meals/{id}/edit' ,'mealController@edit')->middleware(['auth', 'role:admin']);
+Route::PATCH('/meals/{id}' ,'mealController@update')->middleware(['auth', 'role:admin']);
+Route::delete('/meals/{id}'   ,'mealController@destroy')->middleware(['auth', 'role:admin']);
 
 
-// Route::get('/', function () {
-//     return view('welcome');
-// });
+//customer login
+Route::get('/customer/login' ,'customerLoginController@showLoginForm');
+Route::post('/customer/login' ,'customerLoginController@login');
 
+
+Route::get('/profile','profileController@show');
 Auth::routes();
 
-// Route::get('/home', 'HomeController@index')->name('home');
+Route::post('/cart-items', 'CartItemController@store');
+Route::get('/cart-items/checkout', 'CartItemController@checkout');
+
+Route::get('/cart-items', 'CartItemController@index');
